Secure urls

+++ backend/locationApp/key/routes.py
@@ -1,22 +1,21 @@
 from flask import Blueprint, render_template, url_for, flash, redirect, request
 from locationApp import db
-from flask_login import current_user
+from flask_login import current_user, login_required
 from locationApp.key.forms import KeyForm
 from locationApp.models import ApiKey
 from secrets import token_hex
-import requests
 
 
 key = Blueprint('key', __name__)
 
 @key.route("/key-list", methods=['POST', 'GET'])
+@login_required
 def key_list():
-#  key = ApiKey.query.filter_by(owner=current_user)
-  keys = ApiKey.query.all()
-#  active = Borrow.query.filter(Borrow.borrowed_to >= datetime.utcnow()).first()
+  keys = ApiKey.query.filter(ApiKey.user_id == current_user.id).all()
   return render_template('key_list.html', keys=keys)
 
 @key.route("/key/new", methods=['POST', 'GET'])
+@login_required
 def key_new():
   form = KeyForm()
   if form.validate_on_submit():
@@ -30,9 +29,13 @@ def key_new():
   return render_template('key_new.html', form=form)
 
 @key.route("/key/<int:key_id>/delete", methods=['POST', 'GET'])
+@login_required
 def key_delete(key_id):
   key = ApiKey.query.get_or_404(key_id)
-  db.session.delete(key)
-  db.session.commit()
-  flash('Your key has been deleted!', 'success')
+  if key.owner == current_user.id:
+    db.session.delete(key)
+    db.session.commit()
+    flash('Your key has been deleted!', 'success')
+  else:
+    flash('Sorry but you are not authorized to modify this key', 'error')
   return redirect(url_for('key.key_list'))
